Refactor landing page structure and remove unused assets
- Created a new landing page component at apps/cms-app/app/(landing)/page.tsx.
- Deleted the old home page component at apps/cms-app/app/page.tsx.
- Removed unused SVG assets from the public directory.
- Added new hero image and testimonials image to the public directory.
- Deleted the landing hero component from shared components as it is no longer used.
